\name{MethTargetedNGS-package}
\alias{MethTargetedNGS-package}
\alias{MethTargetedNGS}
\docType{package}
\title{
Methylation Analysis of Next Generation Sequencing data.
}
\description{
This package helps in visualizing methylation in CpG sites in NGS data for given datasets (normal/tumor) and to identify differentially methylated CpG sites in normal/tumor.
This package to help in perform profile hidden markov modelling of given sequences.   

NOTE: For profile hidden markov model HMMER software is required
}
\details{
\tabular{ll}{
Package: \tab MethTargetedNGS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-20\cr
License: \tab Artistic-2.0\cr
}
Compare methylation status/pattern between samples.

*\code{compare_samples(healthy,tumor)}

Sequence alignment and create methylation pattern

*\code{methAlign(sequence_fasta, ref_seq)}
}
\author{
Muhammad Ahmer Jamil, Prof. Holger Frohlich, Priv.-Doz. Dr. Osman El-Maarri

Maintainer: Muhammad Ahmer Jamil \email{engr.ahmerjamil@gmail.com}
}

\keyword{ Methylation }

